requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.0
pluggy==1.4.0
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
pytest-django==4.8.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
